Use Rack for running the test webserver.
This patch lets the tests be completely self-sufficient.  You can now
just run them with `rspec spec/*spec.rb`.  Previously, the webserver
had to be started out of band.

Additionally this patch fixes an issue where PhantomJS was not
registering an error when it could not reach the test webpage.